requirements.txt
fire==0.7.1
gdown==5.2.1
litellm==1.82.0
pebble==5.2.0
python_dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
timeout_decorator==0.5.0
tqdm==4.67.1
vllm==0.16.0
